refactor(products): split 'superdesk-products' module.
`superdesk-products` becomes:
```
.
├── controllers
│   ├── ProductsConfigController.js
│   └── index.js
├── services
│   ├── ProductsFactory.js
│   └── index.js
├── styles
│   └── products.less
├── views
│   ├── product-test.html
│   ├── products-config-modal.html
│   ├── products.html
│   └── settings.html
└── index.js

4 directories, 10 files
```
