+++ scripts/superdesk-products/controllers/ProductsConfigController.js
@@ -1,102 +1,6 @@
-var app = angular.module('superdesk.products', [
-    'superdesk.users'
-]);
-
-app
-    .config(['superdeskProvider', function(superdesk) {
-        superdesk
-        .activity('/settings/products', {
-                label: gettext('Products'),
-                controller: ProductsConfigController,
-                templateUrl: 'scripts/superdesk-products/views/settings.html',
-                category: superdesk.MENU_SETTINGS,
-                privileges: {products: 1}
-            });
-    }])
-    .config(['apiProvider', function(apiProvider) {
-        apiProvider.api('products', {
-            type: 'http',
-            backend: {
-                rel: 'products'
-            }
-        });
-    }])
-    .controller('ProductsConfigCtrl', ProductsConfigController)
-    .factory('products', ['$q', 'api', 'contentFilters', '$filter',
-        function($q, api, contentFilters, $filter) {
-            /**
-             * Recursivly returns all products
-             *
-             * @return {*}
-             */
-            var _getAllProducts = function(page, products) {
-                page = page || 1;
-                products = products || [];
-
-                return api('products')
-                .query({max_results: 200, page: page})
-                .then(function(result) {
-                    products = products.concat(result._items);
-                    if (result._links.next) {
-                        page++;
-                        return _getAllProducts(page, products);
-                    }
-                    return $filter('sortByName')(products);
-                });
-            };
-
-            var productsService = {
-                products: null,
-                contentFilters: null,
-                productLookup: {},
-                fetchProducts: function() {
-                    var self = this;
-
-                    return _getAllProducts()
-                    .then(function(result) {
-                        self.products = result;
-                        _.each(result._items, function(product) {
-                            self.productLookup[product._id] = product;
-                        });
-                    });
-                },
-                fetchAllProducts: function() {
-                    return _getAllProducts();
-                },
-                fetchContentFilters: function() {
-                    var self = this;
-
-                    return contentFilters.getAllContentFilters().then(function(filters) {
-                        self.contentFilters = $filter('sortByName')(filters);
-                    });
-                },
-                testProducts: function(diff) {
-                    return api.save('product_tests', {}, diff);
-                },
-                initialize: function() {
-                    return this.fetchProducts()
-                    .then(angular.bind(this, this.fetchContentFilters));
-                }
-            };
-            return productsService;
-        }])
-    .directive('sdProductsConfig', function() {
-        return {
-            controller: ProductsConfigController
-        };
-    })
-    .directive('sdProductsConfigModal', function() {
-        return {
-            templateUrl: 'scripts/superdesk-products/views/products-config-modal.html',
-            link: function(scope, elem, attrs, ctrl) {
-
-            }
-        };
-    });
-
 ProductsConfigController.$inject = ['$scope', 'gettext', 'notify', 'api', 'products', 'modal',
-'subscribersService', 'metadata', '$filter'];
-function ProductsConfigController ($scope, gettext, notify, api, products, modal,
+    'subscribersService', 'metadata', '$filter'];
+export function ProductsConfigController ($scope, gettext, notify, api, products, modal,
     subscribersService, metadata, $filter) {
 
     $scope.testLookup = {};
@@ -222,5 +126,3 @@ function ProductsConfigController ($scope, gettext, notify, api, products, modal
         });
     };
 }
-
-export default app;
