+++ scripts/superdesk-products/controllers/index.js
@@ -1,0 +1,1 @@
+export { ProductsConfigController } from './ProductsConfigController';
