+++ scripts/superdesk-products/services/index.js
@@ -1,0 +1,1 @@
+export { ProductsFactory} from './ProductsFactory';
